+++ packages/cli/src/cli/utils/config.ts
@@ -1,7 +1,7 @@
 import _ from "lodash";
 import fs from "fs";
 import path from "path";
-import { I18nConfig, parseI18nConfig } from "@lingo.dev/~spec";
+import { I18nConfig, parseI18nConfig } from "@lingo.dev/_spec";
 
 export function getConfig(resave = true): I18nConfig | null {
   const configFilePath = _getConfigFilePath();