+++ packages/sdk/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @lingo.dev/~sdk
+# @lingo.dev/\_sdk
 
 ## 0.7.12
 
@@ -7,7 +7,7 @@
 - [`6281dbd`](https://github.com/lingodotdev/lingo.dev/commit/6281dbd96bd5cfe54f194a6a1d055c8255a250de) Thanks [@maxprilutskiy](https://github.com/maxprilutskiy)! - fix sdk/spec exported types
 
 - Updated dependencies [[`6281dbd`](https://github.com/lingodotdev/lingo.dev/commit/6281dbd96bd5cfe54f194a6a1d055c8255a250de)]:
-  - @lingo.dev/~spec@0.24.2
+  - @lingo.dev/_spec@0.24.2
 
 ## 0.7.11
 
@@ -16,7 +16,7 @@
 - [#419](https://github.com/lingodotdev/lingo.dev/pull/419) [`a45feb1`](https://github.com/lingodotdev/lingo.dev/commit/a45feb1d747f8fa32c42c1726953a04c174e754a) Thanks [@maxprilutskiy](https://github.com/maxprilutskiy)! - Replexica is now Lingo.dev! 🎉
 
 - Updated dependencies [[`a45feb1`](https://github.com/lingodotdev/lingo.dev/commit/a45feb1d747f8fa32c42c1726953a04c174e754a)]:
-  - @lingo.dev/~spec@0.24.1
+  - @lingo.dev/_spec@0.24.1
 
 ## 0.7.10